Moved all types into a single file.
