+++ server/src/model/QueryHandlers.model.ts
@@ -3,104 +3,7 @@ import { connectToDB } from "../db";
 import { chatMembers, messages, chats, user, privateChats, privateMessages } from "../schema";
 import { UserSchema } from "./Auth.model";
 import { sql, desc, eq, asc, and, or, inArray, ne, } from "drizzle-orm";
-
-export type SQLErrorType = {
-  error: boolean,
-  reason: string
-}
-
-export interface UserBase {
-  pk_user_id: number;
-  name: string | null;
-  email: string | null;
-  created_at: Date;
-  updated_at?: Date;
-}
-
-export interface MessageBase {
-  id: number;
-  fk_user_id: number;
-  sent_at: Date;
-  message_text: string | null;
-}
-
-export interface ChatBase {
-  pk_chats_id: number;
-  chat_name: string | null;
-  createdAt: Date;
-}
-
-export interface PrivateChatBase {
-  pk_private_chat_id: number;
-  sender_id: number;
-  recipient_id: number;
-  created_at: Date;
-}
-
-export interface PrivateMessageBase {
-  id: number;
-  fk_private_chat_id: number;
-  fk_user_id: number;
-  message_text: string | null;
-  sent_at: Date;
-  image_file: Buffer | string | null;
-  image_name: string | null;
-}
-
-
-export interface PrivateMessageType {
-  private_chat: PrivateChatBase | null;
-  private_messages: PrivateMessageBase | null;
-  chat_user: UserBase | null;
-}
-
-
-// Extended Types
-export interface PrivateMessageType {
-  private_chat: PrivateChatBase | null;
-  private_messages: PrivateMessageBase | null;
-  chat_user: UserBase | null;
-}
-export interface MessageType extends MessageBase {
-  fk_chat_id: number;
-}
-
-export interface ChatType extends ChatBase {
-  chat_members?: ChatMembersType[];
-  messages?: MessageType;
-}
-
-
-export interface ChatMembersType {
-  id: number;
-  fk_chat_id: number;
-  fk_user_id: number;
-  added_at: Date;
-}
-
-export type ChatListType = {
-  chat_members: ChatMembersType
-  chats: ChatType
-  messages?: MessageType
-}[]
-
-export interface DBUserInterface extends UserBase {
-  password: string | null;
-}
-
-
-export interface PrivateChatResult {
-  private_chat: PrivateChatBase;
-  chat_user: UserBase;
-  private_messages: PrivateMessageBase;
-  recipient?: UserBase;
-}
-
-export interface TypedMessage {
-  chats: ChatBase;
-  messages: MessageBase | null;
-  chat_user: UserBase | null;
-}
+import { ChatListType, ChatType, PrivateMessageType, PrivateChatResult, TypedMessage, MessageType, SQLErrorType } from "../types";
 
 export class QueryHandlers extends UserSchema {
   db: NodePgDatabase;
